:recycle:  useProfileFetch hook
